Changes made based on review 13.9
