+++ frontend/src/components/project/ProjectCommentsContent.js
@@ -9,7 +9,6 @@ import getTexts from "../../../public/texts/texts.js";
 import CommentInput from "../communication/CommentInput.js";
 import UserContext from "../context/UserContext.js";
 import Posts from "./../communication/Posts.js";
-import ProjectSliderWithTitle from "./ProjectSliderWithTitle.js";
 
 const useStyles = makeStyles((theme) => {
   return {
@@ -20,14 +19,7 @@ const useStyles = makeStyles((theme) => {
   };
 });
 
-export default function CommentsContent({
-  user,
-  project,
-  token,
-  setCurComments,
-  screenSize,
-  similarProjects,
-}) {
+export default function CommentsContent({ user, project, token, setCurComments }) {
   const classes = useStyles();
   const { locale } = useContext(UserContext);
   const texts = getTexts({ page: "project", locale: locale });
@@ -103,15 +95,6 @@ export default function CommentsContent({
           onSendComment={onSendComment}
           onDeletePost={onDeleteComment}
         />
-      )}
-
-      {screenSize.belowSmall && (
-        <ProjectSliderWithTitle
-          title={texts.you_may_also_like_these_projects}
-          similarProjects={similarProjects}
-        >
-          {" "}
-        </ProjectSliderWithTitle>
       )}
     </div>
   );
